+++ ReleaseNotes.md
@@ -1,4 +1,4 @@
 # Bluedot ReactNative Minimal Integration
 
-- Using [Bluedot plugin 3.2.2](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@>=0.76.6).
-- Support React Native 0.76.6
+- Using [Bluedot plugin 3.3.1](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@>=0.77.0).
+- Support React Native 0.77.0.